+++ src/decorators/index.ts
@@ -1,0 +1,4 @@
+export * from './controller'
+export * from './httpMethod'
+export * from './handlerParam'
+export * from './inject'
