+++ src/exceptions/BadGatewayException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class BadGatewayException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.BAD_GATEWAY)
+    super(message, HttpStatus.BAD_GATEWAY)
   }
 }
