+++ src/exceptions/BadRequestException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class BadRequestException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.BAD_REQUEST)
+    super(message, HttpStatus.BAD_REQUEST)
   }
 }
