+++ src/exceptions/ForbiddenException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class ForbiddenException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.FORBIDDEN)
+    super(message, HttpStatus.FORBIDDEN)
   }
 }
