+++ src/exceptions/GoneException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class GoneException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.GONE)
+    super(message, HttpStatus.GONE)
   }
 }
