+++ src/exceptions/HttpException.ts
@@ -1,9 +1,9 @@
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class HttpException extends Error {
   constructor(
     public readonly message: string,
-    public readonly status: number = HttpStatusEnum.INTERNAL_SERVER_ERROR
+    public readonly status: number = HttpStatus.INTERNAL_SERVER_ERROR
   ) {
     super(message)
   }
