+++ src/exceptions/ImATeapotException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class ImATeapotException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.I_AM_A_TEAPOT)
+    super(message, HttpStatus.I_AM_A_TEAPOT)
   }
 }
