+++ src/exceptions/InternalServerErrorException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class InternalServerErrorException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.INTERNAL_SERVER_ERROR)
+    super(message, HttpStatus.INTERNAL_SERVER_ERROR)
   }
 }
