+++ src/exceptions/NotAcceptableException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class NotAcceptableException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.NOT_ACCEPTABLE)
+    super(message, HttpStatus.NOT_ACCEPTABLE)
   }
 }
