+++ src/exceptions/PayloadTooLargeException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class PayloadTooLargeException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.PAYLOAD_TOO_LARGE)
+    super(message, HttpStatus.PAYLOAD_TOO_LARGE)
   }
 }
