+++ src/exceptions/ServiceUnavailableException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class ServiceUnavailableException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.SERVICE_UNAVAILABLE)
+    super(message, HttpStatus.SERVICE_UNAVAILABLE)
   }
 }
