+++ src/exceptions/UnauthorizedException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class UnauthorizedException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.UNAUTHORIZED)
+    super(message, HttpStatus.UNAUTHORIZED)
   }
 }
