+++ src/exceptions/UnprocessableEntityException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class UnprocessableEntityException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.UNPROCESSABLE_ENTITY)
+    super(message, HttpStatus.UNPROCESSABLE_ENTITY)
   }
 }
