+++ src/exceptions/UnsupportedMediaTypeException.ts
@@ -1,8 +1,8 @@
 import { HttpException } from './HttpException'
-import { HttpStatusEnum } from '../HttpStatusEnum'
+import { HttpStatus } from '../consts'
 
 export class UnsupportedMediaTypeException extends HttpException {
   constructor(public readonly message: string) {
-    super(message, HttpStatusEnum.UNSUPPORTED_MEDIA_TYPE)
+    super(message, HttpStatus.UNSUPPORTED_MEDIA_TYPE)
   }
 }
