+++ src/index.ts
@@ -1,10 +1,7 @@
 import { tachijs } from './tachijs'
 export * from './tachijs'
-export * from './controller'
-export * from './httpMethods'
-export * from './handlerParam'
+export * from './decorators'
 export * from './results'
-export * from './inject'
 export * from './BaseController'
 
 export default tachijs
